requirements.txt
dlt[postgres]>=1.14.1
requests
python-dotenv
